+++ weldx/asdf/cli/welding_schema.py
@@ -1,11 +1,11 @@
 """single_pass_weld schema."""
 import sys
 from io import BytesIO
-from typing import Optional, Tuple
+from typing import Optional, Tuple, Union
 
 
 def single_pass_weld_example(
-    out_file: str = "single_pass_weld_example.asdf",
+    out_file: Optional[Union[str, BytesIO]] = "single_pass_weld_example.asdf",
 ) -> Optional[Tuple[BytesIO, dict]]:
     """Create ASDF file containing all required fields of the single_pass_weld schema.