+++ weldx/tags/core/transformations/_legacy/__init__.py
@@ -1,5 +1,0 @@
-from .coordinate_system_hierarchy import (
-    CoordinateSystemManagerConverter,
-    CoordinateSystemManagerSubsystemConverter,
-    CoordinateTransformationConverter,
-)